Core implementation done.
